[TASK] Move some validator tests to unit tests
This makes the tests faster.
