+++ Tests/Functional/Validation/UserValidatorTest.php
@@ -9,8 +9,6 @@ use OliverKlee\Onetimeaccount\Validation\UserValidator;
 use TYPO3\CMS\Core\Localization\LanguageService;
 use TYPO3\CMS\Extbase\Utility\LocalizationUtility;
 use TYPO3\CMS\Extbase\Validation\Error;
-use TYPO3\CMS\Extbase\Validation\Validator\AbstractValidator;
-use TYPO3\CMS\Extbase\Validation\Validator\ValidatorInterface;
 use TYPO3\TestingFramework\Core\Functional\FunctionalTestCase;
 
 /**
@@ -64,86 +62,6 @@ final class UserValidatorTest extends FunctionalTestCase
         $this->subject = new UserValidator();
     }
 
-    private function buildFullModel(): FrontendUser
-    {
-        $user = new FrontendUser();
-        $user->setName('Ben Best');
-        $user->setFirstName('Ben');
-        $user->setLastName('Best');
-        $user->setAddress('At the multi-core machine 3');
-        $user->setTelephone('+49 1111 2222222');
-        $user->setEmail('ben@example.com');
-        $user->setTitle('developer');
-        $user->setZip('12345');
-        $user->setCity('Development Hill');
-        $user->setCountry('Buthan');
-        $user->setWww('https://example.com');
-        $user->setCompany('TYPO3 Community');
-        $user->setZone('ABC');
-        $user->setPrivacy(true);
-        $user->setFullSalutation('Yo Ben!');
-        $user->setGender(FrontendUser::GENDER_MALE);
-        $user->setStatus(FrontendUser::STATUS_STUDENT);
-        $user->setComments('Wonderful!');
-        $user->setDateOfBirth(new \DateTime('now'));
-
-        return $user;
-    }
-
-    /**
-     * @test
-     */
-    public function isValidator(): void
-    {
-        self::assertInstanceOf(ValidatorInterface::class, $this->subject);
-        self::assertInstanceOf(AbstractValidator::class, $this->subject);
-    }
-
-    /**
-     * @test
-     */
-    public function validateWithNonUserValueReturnsNoErrors(): void
-    {
-        $result = $this->subject->validate(new \stdClass());
-
-        self::assertFalse($result->hasErrors());
-    }
-
-    /**
-     * @test
-     */
-    public function validateWithEmptyUserForNoShowOrRequiredFieldsSettingReturnsNoErrors(): void
-    {
-        $result = $this->subject->validate(new FrontendUser());
-
-        self::assertFalse($result->hasErrors());
-    }
-
-    /**
-     * @test
-     */
-    public function validateWithEmptyUserForEmptyRequiredAndVisibleFieldsReturnsNoErrors(): void
-    {
-        $this->subject->setSettings(['fieldsToShow' => '', 'requiredFields' => '']);
-
-        $result = $this->subject->validate(new FrontendUser());
-
-        self::assertFalse($result->hasErrors());
-    }
-
-    /**
-     * @test
-     */
-    public function validateWithEmptyUserForNoRequiredFieldsButAllVisibleReturnsNoErrors(): void
-    {
-        $concatenatedFields = \implode(',', self::VALIDATABLE_FIELDS);
-        $this->subject->setSettings(['fieldsToShow' => $concatenatedFields, 'requiredFields' => '']);
-
-        $result = $this->subject->validate(new FrontendUser());
-
-        self::assertFalse($result->hasErrors());
-    }
-
     /**
      * @return array<string, array{0: non-empty-string}>
      */
@@ -190,31 +108,5 @@ final class UserValidatorTest extends FunctionalTestCase
         $result = $this->subject->validate(new FrontendUser());
 
         self::assertTrue($result->hasErrors());
-    }
-
-    /**
-     * @test
-     */
-    public function validateModelWithFullModelForAllFieldsRequiredAndVisibleReturnsNoErrors(): void
-    {
-        $concatenatedFields = \implode(',', self::VALIDATABLE_FIELDS);
-        $this->subject->setSettings(['fieldsToShow' => $concatenatedFields, 'requiredFields' => $concatenatedFields]);
-
-        $result = $this->subject->validate($this->buildFullModel());
-
-        self::assertFalse($result->hasErrors());
-    }
-
-    /**
-     * @test
-     */
-    public function validateModelWithFullModelForNoFieldsRequiredAndAllVisibleReturnsNoErrors(): void
-    {
-        $concatenatedFields = \implode(',', self::VALIDATABLE_FIELDS);
-        $this->subject->setSettings(['fieldsToShow' => $concatenatedFields, 'requiredFields' => '']);
-
-        $result = $this->subject->validate($this->buildFullModel());
-
-        self::assertFalse($result->hasErrors());
     }
 }
